+++ heim-disk/src/sys/windows/bindings/mod.rs
@@ -1,3 +1,11 @@
-pub mod volumes;
-pub mod disks;
-pub mod disks2;
+mod drive;
+mod drives;
+mod volumes;
+mod perf;
+mod drive_type;
+
+pub use crate::os::windows::DriveType;
+pub use self::drive::Drive;
+pub use self::drives::Drives;
+pub use self::volumes::Volumes;
+pub use self::perf::disk_performance;